backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
websockets>=12.0
opencv-python>=4.8.0
mediapipe>=0.10.13
numpy>=1.26.0
pydantic>=2.5.0
requests>=2.31.0